NCS-192 styling and refactoring for PR
