+++ src/controllers/confirm.company.controller.ts
@@ -1,12 +1,11 @@
-import {NextFunction, Request, Response} from "express";
+import { NextFunction, Request, Response } from "express";
 import { Templates } from "../types/template.paths";
-
 import { getCompanyProfile } from "../services/company.profile.service";
 import { CompanyProfile } from "@companieshouse/api-sdk-node/dist/services/company-profile/types";
-import {createConfirmationStatement} from "../services/confirmation.statement.service";
-import {Session} from "@companieshouse/node-session-handler";
-import {PRIVATE_SDK_FEATURE_FLAG} from "../utils/properties";
-import {isActiveFeature} from "../utils/feature.flag";
+import { createConfirmationStatement } from "../services/confirmation.statement.service";
+import { Session } from "@companieshouse/node-session-handler";
+import { PRIVATE_SDK_FEATURE_FLAG } from "../utils/properties";
+import { isActiveFeature } from "../utils/feature.flag";
 
 export const get = async (req: Request, res: Response) => {
   const company: CompanyProfile = await getCompanyProfile(req.query.companyNumber as string);
@@ -15,10 +14,10 @@ export const get = async (req: Request, res: Response) => {
 };
 
 export const post = async (req: Request, res: Response, next: NextFunction) => {
-    if(isActiveFeature(PRIVATE_SDK_FEATURE_FLAG)) {
-        const transactionId: string = "";
-        const session: Session = req.session as Session;
-        await createConfirmationStatement(session, transactionId);
-    }
-    next();
+  if (isActiveFeature(PRIVATE_SDK_FEATURE_FLAG)) {
+    const transactionId: string = "";
+    const session: Session = req.session as Session;
+    await createConfirmationStatement(session, transactionId);
+  }
+  next();
 };
