+++ src/utils/logger.ts
@@ -4,3 +4,8 @@ import ApplicationLogger from "@companieshouse/structured-logging-node/lib/Appli
 const logger: ApplicationLogger = createLogger("Confirmation Statement Web");
 
 export default logger;
+
+export const logAndThrowError = (error: Error) => {
+  logger.error(`${error.message} - ${error.stack}`);
+  throw error;
+};
